requirements.txt
beautifulsoup4
requests
selenium
python-docx
webdriver-manager
pytesseract
Pillow